Refactor report to extract generic part; fix some bugs
