+++ src/agr/templates/report.html.jinja
@@ -6,11 +6,11 @@
   <body>
     <h1>{{ report.name }}</h1>
     {% for chapter in report.chapters -%}
-      {% if chapter.name is not None -%}
+      {% if chapter.name is not none -%}
         <h2>{{ chapter.name }}</h2>
       {% endif -%}
       {% for section in chapter.sections -%}
-        {% if section.name is not None -%}
+        {% if section.name is not none -%}
           <h3>{{ section.name }}</h3>
         {% endif -%}
 
@@ -21,7 +21,7 @@
             {% endif -%}
             {% for column in chapter.columns -%}
               <td>
-                {% if column is not None -%}
+                {% if column is not none -%}
                   <h4>{{ column }}</h4>
                 {% endif -%}
               </td>
@@ -31,32 +31,32 @@
             <tr>
               {% if section.named_rows -%}
                 <td>
-                  {% if row.name is not None -%}
+                  {% if row.name is not none -%}
                     {{ row.name }}
                   {% endif -%}
                 </td>
               {% endif -%}
               {% for column in chapter.columns -%}
                 <td>
-                  {% if row.by_column[column] is None -%}
+                  {% if row.by_column[column] is none -%}
                     <span>(missing)</span>
-                  {% elif section.kind == "image" -%}
+                  {% elif isinstance(row.by_column[column], Image) -%}
                     <img
-                      src="{{ row.by_column[column] }}"
+                      src="{{ row.by_column[column].url }}"
                       title="{{ row.description }}"
                       height="300"
                       width="300"
                     />
-                  {% elif section.kind == "link" -%}
+                  {% elif isinstance(row.by_column[column], Link) -%}
                     <a
-                      href="{{ row.by_column[column] }}"
-                      target="{{ row.by_column[column] }}"
+                      href="{{ row.by_column[column].url }}"
+                      target="{{ row.by_column[column].url }}"
                     >
                       {{ row.name }}
                     </a>
-                  {% elif section.kind == "inline" -%}
+                  {% elif isinstance(row.by_column[column], Inline) -%}
                     <font size="-2">
-                      <pre>{{ row.by_column[column] }}</pre>
+                      <pre>{{ row.by_column[column].content }}</pre>
                     </font>
                   {% endif -%}
                 </td>
