+++ src/agr/util/report.py
@@ -1,30 +1,45 @@
 from dataclasses import dataclass
 from jinja2 import Environment, PackageLoader, select_autoescape
-from typing import Literal, Optional
+from typing import Optional
 
 # HTML reports with multiple columns and multiple sections, defined in a Jinja template.
 # The list of columns is defined per chapter.
 # Sections may be rendered with or without row names.
 
 
+@dataclass
+class Image:
+    url: str
+
+
+@dataclass
+class Link:
+    url: str
+
+
+@dataclass
+class Inline:
+    content: str
+
+
+Target = Image | Link | Inline
+
+
 @dataclass(kw_only=True)
 class Row:
     """The target per column."""
 
-    name: Optional[str] = None
+    name: str
     description: Optional[str] = None
-    by_column: dict[
-        str, Optional[str]
-    ]  # for inline kind this is the file content not the path
+    by_column: dict[str, Optional[Target]]
 
 
 @dataclass(kw_only=True)
 class Section:
-    """A section contains multiple rows all of the same kind."""
+    """A section contains multiple rows with an optional heading."""
 
     name: Optional[str] = None
     named_rows: bool = False
-    kind: Literal["image", "link", "inline"]
     rows: list[Row]
 
 
@@ -43,8 +58,17 @@ class Report:
     chapters: list[Chapter]
 
 
-def render_report(report: Report, out_path: str, template="report.html.jinja"):
-    env = Environment(loader=PackageLoader("agr.util"), autoescape=select_autoescape())
+def render_report(
+    report: Report, out_path: str, module="agr", template="report.html.jinja"
+):
+    """
+    Render template f"{module}/templates/{template}" using `report`.
+    """
+    env = Environment(loader=PackageLoader(module), autoescape=select_autoescape())
+
+    # register extra functions and classes we want to use in the Jinja template
+    env.globals.update(isinstance=isinstance, Image=Image, Link=Link, Inline=Inline)
+
     template = env.get_template(template)
     with open(out_path, "w") as out_f:
         _ = out_f.write(template.render(report=report))
